requirement.txt
argparse
time
os
colorama
pyfiglet
requests
fpdf
paramiko
PyPDF2
python-dotenv